+++ src/main/java/com/example/sync/repository/UsuariosRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuariosRepository extends JpaRepository<Usuarios, Integer> {
+public interface UsuariosRepository extends JpaRepository<Usuarios, Long> {
     Usuarios findByEmail(String email);
 }
